Minor changes to EventDateSection
