+++ ui-components/index.ts
@@ -107,13 +107,13 @@ export * from "./src/page_components/listing/listing_sidebar/events/DownloadLott
 export * from "./src/page_components/listing/listing_sidebar/events/OpenHouseEvent"
 export * from "./src/page_components/listing/listing_sidebar/events/LotteryResultsEvent"
 export * from "./src/page_components/listing/listing_sidebar/events/PublicLotteryEvent"
+export * from "./src/page_components/listing/listing_sidebar/events/EventDateSection"
 
 /* Responsive Wrappers */
 export * from "./src/sections/ResponsiveWrappers"
 
 /* Sections */
 export * from "./src/sections/ContentSection"
-export * from "./src/sections/EventDateSection"
 export * from "./src/sections/FooterSection"
 export * from "./src/sections/InfoCardGrid"
 export * from "./src/sections/ListSection"
